+++ test/observable_test_utils.dart
@@ -2,16 +2,16 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library observable.test.observable_test_utils;
+library change_notifier.test.observable_test_utils;
 
-import 'package:observable/observable.dart';
+import 'package:change_notifier/change_notifier.dart';
 import 'package:test/test.dart';
 
 /// A small method to help readability. Used to cause the next "then" in a chain
 /// to happen in the next microtask:
 ///
 ///     future.then(newMicrotask).then(...)
-Future newMicrotask(_) => Future.value();
+Future<Object?> newMicrotask(Object? _) => Future<Object?>.value();
 
 void expectChanges(List<ChangeRecord> actual, List<ChangeRecord> expected,
     {String? reason}) {
@@ -23,18 +23,21 @@ void expectNotChanges(List<ChangeRecord> actual, ChangeRecords expectedNot,
   expect(actual, isNot(_EqualsMatcher(expectedNot)), reason: reason);
 }
 
-List<ListChangeRecord> getListChangeRecords(
-        List<ListChangeRecord> changes, int index) =>
-    List.from(changes.where((ListChangeRecord c) => c.indexChanged(index)));
+List<ListChangeRecord<E>> getListChangeRecords<E>(
+        List<ListChangeRecord<E>> changes, int index) =>
+    List.from(changes.where((ListChangeRecord<E> c) => c.indexChanged(index)));
 
-List<PropertyChangeRecord> getPropertyChangeRecords(
-        List<ChangeRecord> changes, Symbol property) =>
-    List.from(changes.where(
-        (ChangeRecord c) => c is PropertyChangeRecord && c.name == property));
+List<PropertyChangeRecord<Object?, Symbol>> getPropertyChangeRecords(
+    List<ChangeRecord> changes, Symbol property) {
+  return List.from(changes
+      .whereType<PropertyChangeRecord<Object?, Symbol>>()
+      .where((PropertyChangeRecord<Object?, Symbol> c) => c.name == property));
+}
 
 List<Matcher> changeMatchers(List<ChangeRecord> changes) => changes
-    .map((r) =>
-        r is PropertyChangeRecord ? _PropertyChangeMatcher(r) : equals(r))
+    .map((r) => r is PropertyChangeRecord
+        ? _PropertyChangeMatcher(r as PropertyChangeRecord<Object?, Symbol>)
+        : equals(r))
     .toList();
 
 // Custom equality matcher is required, otherwise expect() infers ChangeRecords
@@ -49,12 +52,12 @@ class _EqualsMatcher<ValueType> extends Matcher {
       description.addDescriptionOf(_expected);
 
   @override
-  bool matches(dynamic item, Map matchState) =>
+  bool matches(dynamic item, Map<dynamic, dynamic> matchState) =>
       item is ChangeRecords && _expected == item;
 }
 
-class _PropertyChangeMatcher<ValueType> extends Matcher {
-  final PropertyChangeRecord<ValueType> _expected;
+class _PropertyChangeMatcher<ValueType, K> extends Matcher {
+  final PropertyChangeRecord<ValueType, K> _expected;
 
   _PropertyChangeMatcher(this._expected);
 
@@ -63,7 +66,7 @@ class _PropertyChangeMatcher<ValueType> extends Matcher {
       description.addDescriptionOf(_expected);
 
   @override
-  bool matches(dynamic other, Map matchState) =>
+  bool matches(dynamic other, Map<dynamic, dynamic> matchState) =>
       identical(_expected, other) ||
       other is PropertyChangeRecord &&
           _expected.runtimeType == other.runtimeType &&